changes ajax call done register
