+++ tms/static/js/register.js
@@ -1,6 +1,11 @@
 
 var registerURL = "/register/"
 
+var showMessage = function (message, color) {
+        var messageElement = $("<div>").text(message).css("color", color);
+        $("#message-container").html(messageElement);
+    };
+
 $("#user_register").on("submit",function(event){
      event.preventDefault()
      var csrfToken = $('input[name="csrfmiddlewaretoken"]').val()
@@ -13,5 +18,13 @@ $("#user_register").on("submit",function(event){
                         "password_confirm":$("#password_confirm").val()}
 
      var redirectURL = "/login_page/"
-     postAjaxCall(registerURL, csrfToken, resultData,redirectURL)
+     var callback = function(response){
+            showMessage(response.success_message, "green");
+                            setTimeout(function () {
+                                window.location.href = redirectURL;
+                            }, 3000);
+     }
+     postAjaxCall(registerURL, csrfToken, resultData,callback)
 })
+
+
