+++ tms/templates/base.html
@@ -36,7 +36,7 @@
             </div>
 
             <div class="navbar-nav ms-lg-4">
-                <a class="nav-item nav-link" id="logout">logout</a>
+                <a class="nav-item nav-link" id="logout" >logout</a>
             </div>
 
             <div class="navbar-nav ms-lg-4">
@@ -47,6 +47,7 @@
         </div>
     </div>
 </nav>
+<script src="{% static 'js/login.js' %}"></script>
 
 {% block content %}
