+++ tms/templates/home.html
@@ -5,7 +5,6 @@
 {% csrf_token %}
 <script src="{% static 'js/base.js' %}"></script>
 <script src="{% static 'js/logout.js' %}"></script>
-<script src="{% static 'js/login.js' %}"></script>
 
 
 {% endblock %}
