+++ tms/tms/settings/base.py
@@ -198,3 +198,4 @@ REST_FRAMEWORK = {
     ),
     "PAGE_SIZE": 10
 }
+
